test(ci): align SFA CP UT coverage metadata
